+++ src/main/java/com/avensys/rts/contactservice/payloadnewresponse/FormSubmissionsResponseDTO.java
@@ -1,6 +1,7 @@
 package com.avensys.rts.contactservice.payloadnewresponse;
 
 import com.fasterxml.jackson.databind.JsonNode;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,10 +10,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class FormSubmissionsResponseDTO {
-    private Integer id;
-    private String formId;
-    private Integer userId;
-    private JsonNode submissionData;
-    private Integer entityId;
-    private String entityType;
+	private Long id;
+	private String formId;
+	private Long userId;
+	private JsonNode submissionData;
+	private Integer entityId;
+	private String entityType;
 }
